add pyspark dtypes

+++ src/synthesized_datasets/_dtypes.py
@@ -38,7 +38,7 @@ _PS_DTYPE_MAP = {
     DType.DOUBLE: st.DoubleType(),
     DType.INTEGER: st.IntegerType(),
     DType.LONG: st.LongType(),
-    DType.NULLABLE_LONG: st.LongType(),
+    DType.NULLABLE_LONG: st.FloatType(),
     DType.STRING: st.StringType(),
     DType.TIMEDELTA: st.LongType(),
 }
